+++ litmus-portal/frontend/public/locales/en/translation.yaml
@@ -1159,7 +1159,7 @@ workflowAgent:
   noAgentSearch: No agents found
 
 workflows:
-  browseWorkflows: Browse workflows
+  runs: Runs
   schedules: Schedules
   templates: Templates
   analytics: Analytics
